refactor(extraction): move shared LLM agent builder
Move the LLM agent construction logic from `extract_metadata` into a
reusable `_llm.build_agent` helper, and use it in the existing metadata
extraction activity.
